refactor files structure
