--- app/src/main/java/com/mahmoudroid/rickmorty/ui/CharacterDetails.kt
@@ -1,7 +1,0 @@
-package com.mahmoudroid.rickmorty.ui
-
-import androidx.fragment.app.Fragment
-import com.mahmoudroid.rickmorty.R
-
-class CharacterDetails  : Fragment(R.layout.details_character){
-}
