+++ app/src/main/java/com/mahmoudroid/rickmorty/ui/details/CharacterDetailsFragment.kt
@@ -1,0 +1,7 @@
+package com.mahmoudroid.rickmorty.ui.details
+
+import androidx.fragment.app.Fragment
+import com.mahmoudroid.rickmorty.R
+
+class CharacterDetailsFragment  : Fragment(R.layout.details_character){
+}
