+++ app/src/main/java/com/mahmoudroid/rickmorty/ui/home/MainActivity.kt
@@ -1,19 +1,11 @@
-package com.mahmoudroid.rickmorty.ui
+package com.mahmoudroid.rickmorty.ui.home
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
-import androidx.recyclerview.widget.DividerItemDecoration
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.mahmoudroid.rickmorty.R
-import com.mahmoudroid.rickmorty.adapter.CharactersAdapter
 import com.mahmoudroid.rickmorty.databinding.ActivityMainBinding
-import com.mahmoudroid.rickmorty.viewmodel.MainActivityViewModel
-import kotlinx.coroutines.flow.collectLatest
 
 
 private const val TAG = "####MainActivity"
